backend/requirements.txt
cryptography==35.0.0
Django==3.2.8
django-cors-headers==3.10.0
django-random-id-model==0.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
django-environ
